creating mixup file
